Move pop to utils
